refactor: 전역 상수 폴더 세팅 및 이동
